fix(hub-discussions): repackage types into a single file
affects: @esri/hub-discussions
