+++ packages/discussions/src/channels/index.ts
@@ -1,2 +1,1 @@
 export * from "./channels";
-export * from "./types";
