+++ packages/discussions/src/index.ts
@@ -4,7 +4,6 @@
 export * from "./posts";
 export * from "./channels";
 export * from "./reactions";
-export * from "./types";
 
 export * from "./utils/posts";
 export * from "./utils/channels/";
