+++ packages/discussions/src/posts/index.ts
@@ -1,2 +1,1 @@
 export * from "./posts";
-export * from "./types";
