+++ packages/discussions/src/reactions/index.ts
@@ -1,2 +1,1 @@
 export * from "./reactions";
-export * from "./types";
