+++ packages/discussions/src/utils/channels/can-post-to-channel.ts
@@ -4,8 +4,10 @@ import {
   IAclPermission,
   IChannel,
   IChannelAcl,
-} from "../../channels";
-import { IDiscussionsUser, Role, SharingAccess } from "../../shared/types";
+  IDiscussionsUser,
+  Role,
+  SharingAccess,
+} from "../../types";
 import { CANNOT_DISCUSS } from "../constants";
 
 const ALLOWED_GROUP_ROLES = Object.freeze(["owner", "admin", "member"]);
