+++ packages/discussions/src/utils/channels/index.ts
@@ -1,7 +1,6 @@
 import { IUser } from "@esri/arcgis-rest-auth";
 import { GroupMembership } from "@esri/arcgis-rest-portal";
-import { IChannel } from "../../channels";
-import { IPlatformSharing } from "../../shared/types";
+import { IChannel, IPlatformSharing } from "../../types";
 import { isOrgAdmin, reduceByGroupMembership } from "../platform";
 
 function intersectGroups(
